Add ID3 tag parsing, songs stored in linked list, extra audio api functions to control playback.
Huge update - we use the TagLib library now to parse through ID3 files.
This currently limits our accepted filetype to .mp3's though. The UI
should just pass in currently selected ITEM's to loadSound() and the
backend will take care of traversing the linked list and grabbing the
correct metadata. This should make it easier for the UI to express
itself without worrying about implementation specifics of the audio
files.